+++ packages/orq-rc/RELEASES.md
@@ -9679,3 +9679,13 @@ Based on:
 - [python v4.12.0-rc.56] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.56] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.56 - packages/orq-rc
+
+## 2026-07-22 22:35:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.0] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.0 - packages/orq-rc